requirements.txt
python-telegram-bot==20.7
pymongo==4.6.1
python-dotenv==1.0.0
flask==3.0.0
qrcode==7.4.2
pillow==10.1.0
requests==2.31.0
pytz==2023.3
aiohttp==3.9.1
motor==3.3.2
redis==5.0.1
cryptography==41.0.7
